move type selector cards component to a separate file
